requirements.txt
streamlit>=1.30.0
pyautogen>=0.2.0
neo4j>=5.14.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pandas
plotly
watchdog
